requirements.txt
torch>=1.1
torchvision>=0.3
numpy
opencv-python
tqdm>=4.36.0
matplotlib
plotly
scipy
h5py
pycolmap>=0.3.0
kornia>=0.6.7
gdown
scikit-learn
plyfile
PyYAML==6.0